No longer crashes on incorrect dates entered
When a date is entered (as a string), a method is run that tries to create a Date object with that string - if it's unable to, it means the user entered an invalid date, and so the program retrieves the current date and uses that instead. This also helped to remove some code repetition.